% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatrix.R
\name{h_matrix}
\alias{h_matrix}
\title{h_matrix}
\usage{
h_matrix(lr.model)
}
\arguments{
\item{lr.model}{Take an R object, returned by \link[Linear3]{lr}}
}
\value{
A checklist that includes the following:
\itemize{
  \item{maximum hi} - {Which entity has the maximum value of the leverage}
  \item{hi>2*mean(H)} - {Which objects are outliers in the X-space, and their leverage values}
}
}
\description{
Finds the index of the maximum leverage value
}
\examples{
##Annette Dobson (1990) "An Introduction to Generalized Linear Models".
##Page 9: Plant Weight Data.
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
model <- lr(ctl~trt, mtcars)
hat_matrix(model)
hat_matrix(model)[1]
hat_matrix(model)[2]


}
