% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lr.R
\name{lr}
\alias{lr}
\title{lr}
\usage{
lr(formula, data, include.intercept = TRUE, predict = NULL, na.action = "omit")
}
\arguments{
\item{formula}{An object of class "formula" which is the model to fit the regression}

\item{data}{Data frame to perform the linear regression on}

\item{include.intercept}{If the model should be fitted with an intercept, include.intercept = TRUE; if model should be fitted without an intercept,
then include.intercept = FALSE. The default setting for include.intercept is TRUE.}

\item{predict}{The default parameter is set to NULL. If you want to use the current model for prediction, enter an n by p matrix (or
object coercible by as.matrix to a matrix) to get predicted values. n is the number of prediction desired and p is the number of covariates included in the model.}

\item{na.action}{Character input, which determines how the regression model should handle missing data. Options include 'omit' (remove rows with missing values), 'fail'(stop regression), or 'impute' (replace missing values with column mean). Defaults to omit if excluded from call.}
}
\value{
lr doesn't explicitly return anything unless you extract the value with $ followed by the name of the desired output.
The returned output is a list containing at least the following:
\describe{
 \item{coefficients}{a named coefficients vector}
 \item{residuals}{Residuals (i.e. response values minus fitted values)}
 \item{standardized_res}{the standardized residuals}
 \item{studentized_res}{the internally studentized residuals}
 \item{ex_stud_res}{the externally studentized residuals}
 \item{fitted.values}{the predicted value}
 \item{sigma}{the residual standard error}
 \item{hat_matrix}{the matrix maps the vector of response values (dependent variable values) to the vector of fitted values (or predicted values)}
 \item{leverage}{obtain leverage}
 \item{df}{degrees of freedom}
 \item{rank}{the number of betas}
 \item{coeff_summary}{mimic the results using summary(lm()) including estimates the beta coefficients, standard errors, t values, and p-value}
 \item{R_squared}{the proportion of the variation in the dependent variable explained by the independent variable(s)}
 \item{adj_R_squared}{a penalized version of R_squared}
 \item{CI}{95\% confidence interval of estimates (i.e. coefficients)}
 \item{Fstatistic}{give the overall F statistic and its corresponding degrees of freedom of numerator and denominator}
 \item{p_value_F_test}{p-value for overall F test}
 \item{predicted}{give the predicted value using the current fitted model}
}
}
\description{
A fitted linear regression model was used
}
\examples{
data(mtcars)
attach(mtcars)
lr(mpg ~ cyl + wt , mtcars)$coefficients ## Obtain beta coefficient estimates
lr(mpg ~ cyl + wt, mtcars)$coeff_summary ## Obtain summary of beta coefficients
lr(mpg ~ cyl + wt, mtcars)$sigma ## Obtain residual standard error
lr(mpg ~ cyl + wt, mtcars)$CI ## Obtain 95\% confidence interval
lr(mpg ~ cyl + wt, mtcars, include.intercept = FALSE) ## omitting intercept
lr(mpg ~ cyl + wt, mtcars, include.intercept = FALSE)$df ## Extract degrees of freedom when fitting a model without an intercept
lr(mpg ~ cyl + wt, mtcars, predict = matrix(c(mean(mtcars$mpg), mean(mtcars$wt)), 1, 2))$predicted
detach(mtcars)

}
