% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_influence.R
\name{plotdffits}
\alias{plotdffits}
\title{Plot Influential Outliers}
\usage{
\itemize{
  \item{plotdffit(lr.model)} - {Plot a DFFITS graph and indicate the subject of the outlier}
  \item{plotcd(lr.model)}- {Plot a Cook's Distance graph and indicate the subject of the outlier}
  \item{plotCVR(lr.model)} - {Plot a COVRATIO graph }
}
}
\arguments{
\item{lr.model}{Take an R object, returned by \link[Linear3]{lr}}
}
\description{
This set of plotting functions is used to plot
DFFITS, Cook's Distance, and COVRATIO separately
}
\examples{
require(ggplot2)
data(mtcars)
attach(mtcars)
model <- lr(mpg~cyl + wt, mtcars)
plotdffits(model)
plotcd(model)
plotCVR(model)

}
