% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{ANOVA}
\alias{ANOVA}
\title{ANOVA}
\usage{
ANOVA(formula, data, type, na.action = "omit")
}
\arguments{
\item{formula}{An object of class "formula": a symbolic description of the model to be fitted. A typical model is the form outcome ~ covariates
where the outcome is a numeric response vector (which people usually denote as Y in statistical formulas) and the covariates are predictors of response.}

\item{data}{A data frame (or an object convertible to a data frame by as.data.frame) containing the variables in the model.}

\item{type}{The type of sum of squares to obtain. Use "Sequential" for sequential sums of squares and "Partial" for partial sums of squares.}

\item{na.action}{Character input, which determines how the regression model should handle missing data. Options include 'omit' (remove rows with missing values), 'fail'(stop regression), or 'impute' (replace missing values with column mean). Defaults to omit if excluded from call.}
}
\value{
ANOVA returns anova table in a data.frame.
}
\description{
ANOVA is used to obtain an analysis of variance table for a fitted model. It produces the same results as anova(lm( )) for sequential sums of squares
and car::Anova(lm( ), type="III") for partial sums of squares.
}
\examples{

ANOVA(mpg ~ cyl + wt + qsec + disp, mtcars, type = "Partial") ## Get partial SS
ANOVA(mpg ~ cyl + wt + qsec + disp, mtcars, type = "Sequential") ## Get sequential SS
ANOVA(mpg ~ cyl + wt + qsec + disp, mtcars, type = "Sequential")["F value"] ## Extract F statistics from ANOVA table

}
