% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmatrix.R
\name{h_matrix}
\alias{h_matrix}
\title{h_matrix}
\usage{
h_matrix(lr.model)
}
\arguments{
\item{lr.model}{Take an R object, returned by \link[Linear3]{lr}}
}
\value{
A checklist that includes the following:
\itemize{
  \item{maximum hi} - {Which entity has the maximum value of the leverage}
  \item{hi>2*mean(H)} - {Which objects are outliers in the X-space, and their leverage values}
}
}
\description{
Finds the index of the maximum leverage value
}
\examples{
data(mtcars)
attach(mtcars)
model <- lr(mpg ~ cyl + wt, mtcars)
h_matrix(model)
h_matrix(model)[1]
h_matrix(model)[2]
detach(mtcars)

}
