% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_influence.R
\name{plotdffits}
\alias{plotdffits}
\title{Plot Influential Outliers}
\usage{
\itemize{
  \item{plotdffit(lm.model)} - {Plot a DFFITS graph and indicate the subject of the outlier}
  \item{plotcd(lm.model)}- {Plot a Cook's Distance graph and indicate the subject of the outlier}
}
}
\arguments{
\item{lm.model}{Take an R object, returned by \link[stats]{lm}}
}
\description{
This set of plotting functions is used to plot
DFFITS, Cook's Distance separately
}
\examples{
require(ggplot2)
data(mtcars)
attach(mtcars)
model <- lm(mpg~cyl + wt, mtcars)
plotdffits(model)
plotcd(model)

}
