% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_influence.R
\name{plotdffits}
\alias{plotdffits}
\title{Plot Influential Outliers}
\usage{
\itemize{
  \item{plotdffit(lr.model)} - {Plot a DFFITS graph and indicate the subject of the outlier}
  \item{plotcd(lr.model)}- {Plot a Cook's Distance graph and indicate the subject of the outlier}
  \item{plotCVR(lr.model)} - {Plot a COVRATIO graph }
}
}
\arguments{
\item{lr.model}{Take an R object, returned by \link[Linear3]{lr}}
}
\description{
This set of plotting functions is used to plot
DFFITS, Cook's Distance, and COVRATIO separately
}
\examples{
require(ggplot2)
##Annette Dobson (1990) "An Introduction to Generalized Linear Models".
##Page 9: Plant Weight Data.
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
model <- lr(ctl~trt, mtcars)
plotdffits(model)
plotcd(model)
plotCVR(model)

}
