% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_influence.R
\name{plotdffits}
\alias{plotdffits}
\title{Plot Influential Outliers}
\usage{
plotdffits(lm.model)
}
\arguments{
\item{lm.model}{Take an R object, returned by \link[stats]{lm}}
}
\description{
This set of plotting functions is used to plot DFFITS, Cook's Distance separately
}
\examples{
library(ggplot2)
model <- lm(mpg~cyl + wt + qsec + disp, mtcars)
plotdffits(model)
plotcd(model)

}
