% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outliers_influence.R
\name{plotoutliers}
\alias{plotoutliers}
\title{Plot Influential Outliers}
\usage{
plotoutliers(lm.model, option = c("dffits", "cd"))
}
\arguments{
\item{lm.model}{Take an R object, returned by \link[stats]{lm}}
}
\description{
This set of plotting functions is used to plot DFFITS, Cook's Distance separately
}
\examples{
library(ggplot2)
model <- lm(mpg~cyl + wt + qsec + disp, mtcars)
plotoutliers(model,option = "dffits")
plotoutliers(model,option = "cd")

}
